+++ source/Patterns/Composite/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 
 namespace Composite
 {
@@ -27,62 +26,6 @@ namespace Composite
             }
 
             root.WriteAllNames();
-        }
-    }
-
-    internal abstract class Component
-    {
-        public string Name {get; set;}
-
-        public Component(string name)
-        {
-            Name = name;
-        }
-        
-        public abstract void WriteAllNames(int depth=0);
-
-    }
-
-    class Composite : Component
-    {
-        private List<Component> _children  = new List<Component>();
-        public Composite(string name) : base(name) {}
-        public Component Add(Component componnet)
-        {
-            _children.Add(componnet);
-            return componnet;
-        }
-
-        public void Remove(Component componnet)
-        {
-            _children.Remove(componnet);
-        }
-
-        public override void WriteAllNames(int depth = 0)
-        {
-            System.Console.WriteLine($"{new string('-', depth)} {Name}");
-
-            var stack = new Queue<Composite>();
-
-            foreach (var child in _children)
-            {
-                if(child is Leaf) child.WriteAllNames(depth+1);
-                else if(child is Composite) stack.Enqueue((Composite) child);
-            }
-
-            while (stack.Count>0)
-            {
-                stack.Dequeue().WriteAllNames(depth+1);
-            }
-        }
-    }
-
-    class Leaf : Component
-    {
-        public Leaf(string name) : base(name){}
-        public override void WriteAllNames(int depth = 0)
-        {
-            System.Console.WriteLine($"{new string('-', depth )} {Name}");
         }
     }
